+++ src/Umbraco.Tests/Persistence/Repositories/ContentRepositoryTest.cs
@@ -330,7 +330,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IContent>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "Name", Direction.Ascending);
 
                 // Assert
@@ -351,7 +351,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IContent>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 1, 1, out totalRecords, "Name", Direction.Ascending);
 
                 // Assert
@@ -372,7 +372,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IContent>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 2, out totalRecords, "Name", Direction.Ascending);
 
                 // Assert
@@ -393,7 +393,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IContent>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "Name", Direction.Descending);
 
                 // Assert
@@ -414,7 +414,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IContent>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "Name", Direction.Ascending, "Page 2");
 
                 // Assert
@@ -435,7 +435,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IContent>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "Name", Direction.Ascending, "Page");
 
                 // Assert
