+++ src/Umbraco.Tests/Persistence/Repositories/MediaRepositoryTest.cs
@@ -249,7 +249,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Ascending);
 
                 // Assert
@@ -270,7 +270,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 1, 1, out totalRecords, "SortOrder", Direction.Ascending);
 
                 // Assert
@@ -291,7 +291,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 2, out totalRecords, "SortOrder", Direction.Ascending);
 
                 // Assert
@@ -312,7 +312,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Descending);
 
                 // Assert
@@ -333,7 +333,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "Name", Direction.Ascending);
 
                 // Assert
@@ -354,7 +354,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Ascending, "File");
 
                 // Assert
@@ -375,7 +375,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Ascending, "Test");
 
                 // Assert
